refactor(calendar): extract API to build event instance from master event and recurrenceDate
